[Fix] Entity 클래스 내 updatedAt, createdAt 중복 메소드 BaseEntity로 빼서 상속
